requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
promise==2.3
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
text-unidecode==1.3
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
